Demonstrate build "improvements" disabling AUTOMOC/AUTOUIC
Disabling AUTOMOC/AUTOUC
and using the low level qt_wrap_ui/cpp,
results in build improvements,
because all the dependencies are correctly mapped.

Additionally, targets for each generated file are produced,
which avoids build bottlenecks on the combined mocs compilation,
which improves build times in large projects.

This "bottleneck" could be the "pause" which was hinted at here:
https://discourse.cmake.org/t/generate-slow-pauses-for-30-seconds-starting-with-3-21/3901